test(dataset): organize the crowded tests/dataset dir into concern subfolders
Move the 56 top-level test modules into eight new concern packages, and
fold the five collection-related files into the existing collection/
subpackage, mirroring the existing cog/collection/ops layout:

- unit/       test_unit_* + clustering + engines
- plot/       plotting tests (+ _plot_helpers unit tests)
- io/         read/write/bytes/pickle/archive/window/reads/create_empty/from_band_files
- spatial/    dataset_spatial/bbox_crop/merge/resample/warped_view/georef/transform/masks/overviews
- analysis/   vectorize/contour/proximity/sieve/sample/terrain_rgb
- create/     dataset_create/dataset_nodata
- stac/       stac/to_stac_item/from_stac_grid
- remote/     wcs/ogc_coverages (+ the wcs_mock_server helper)
- collection/ (existing) <- dataset_collection(_unit)/from_point/lazy_pipelines/to_cog_stack

conftest.py stays at tests/dataset/ so its fixtures inherit into every
subpackage; each new folder gets an __init__.py. Repoint test_wcs's
wcs_mock_server import to tests.dataset.remote.wcs_mock_server; bump
test_plot_cleanups' repo-root anchor parents[2]->parents[3] for the added
depth. Rename test_datacube_cog.py -> collection/test_to_cog_stack.py
(datacube is the outdated name for DatasetCollection). Pure relocation:
2207 tests collect unchanged, no bodies edited.
