+++ tests/dataset/analysis/test_dataset_vectorize.py
