+++ tests/dataset/analysis/test_proximity.py
