+++ tests/dataset/analysis/test_terrain_rgb.py
