+++ tests/dataset/collection/test_dataset_lazy_pipelines.py
