+++ tests/dataset/collection/test_from_point.py
