+++ tests/dataset/collection/test_to_cog_stack.py
