+++ tests/dataset/io/test_read_array_masked.py
