+++ tests/dataset/io/test_write_array_window.py
