+++ tests/dataset/plot/test_plot_color.py
