+++ tests/dataset/plot/test_plot_resolution.py
