+++ tests/dataset/remote/test_ogc_coverages.py
