+++ tests/dataset/spatial/test_bbox_crop_read_array.py
