+++ tests/dataset/spatial/test_dataset_spatial.py
