+++ tests/dataset/spatial/test_georef.py
