+++ tests/dataset/spatial/test_masks.py
