+++ tests/dataset/spatial/test_merge.py
