+++ tests/dataset/spatial/test_resampling_methods.py
