+++ tests/dataset/spatial/test_warped_view.py
